Silence various compiler warnings on non-MULE.
src/ChangeLog addition:

2017-02-11  Aidan Kehoe  <kehoea@parhasard.net>

	* console-stream.c (stream_text_width):
	* data.c (build_fixnum_to_char_map):
	* data.c (vars_of_data):
	* doprnt.c (fixnum_to_string):
	* doprnt.c (bignum_to_string_1):
	* doprnt.c (parse_doprnt_spec):
	* event-xlike-inc.c (endif):
	* print.c (Fwrite_sequence):
	* print.c (print_internal):
	* search.c (boyer_moore):
	* tests.c (Ftest_character_tell):
	* text.c:
	* text.c (ibyte_string_displayed_columns):

+++ src/tests.c
@@ -654,7 +654,8 @@ REASON is nil or a string describing the failure (not required).
   CHARACTER_TELL_ASSERT (Lstream_character_tell (XLSTREAM (stream))
                          /* This should be equal to sizeof (shortbuf) on
                             non-mule. */
-                         == sizeof (shortbuf) - !(byte_ascii_p (0xff)),
+                         == (ssize_t) (sizeof (shortbuf))
+                         - !(byte_ascii_p (0xff)),
                          "character tell with short read, no-conversion-unix",
                          "short read character tell failed");
